objcopy: provide cache for gnu_debuglink's target CRC v2
Currently objcopy --add-gnu-debuglink=foo.a.dbg foo.a will calculate the
CRC32 of target debug file for each entry inside the archive. The
archive is not expected to change ever and in particular not while
adding the .gnu_debuglink section, so use a cache to avoid recalculating
it over and over again.

I tested it on a sample foo.a:

$ objcopy --only-keep-debug foo.a foo.a.dbg ; objcopy --strip-debug foo.a
$ du -b foo.a foo.a.dbg
14311136	foo.a
142039500	foo.a.dbg
$ ar t foo.a | wc -l
298
$ time objcopy-old --add-gnu-debuglink=foo.a.dbg foo.a
80,26s user 4,87s system 99% cpu 1:25,30 total
$ sha1sum foo.a foo.a.dbg
4ee33b908a960bb2143e9ce7f5c4c1213aa2bdc1  foo.a
19b4b901a0110327a6a234f7a91d387f09ee3a03  foo.a.dbg
$ # repeat the setup
$ time objcopy-new --add-gnu-debuglink=foo.a.dbg foo.a
0,33s user 0,07s system 99% cpu 0,408 total
$ sha1sum foo.a foo.a.dbg
4ee33b908a960bb2143e9ce7f5c4c1213aa2bdc1  foo.a
19b4b901a0110327a6a234f7a91d387f09ee3a03  foo.a.dbg

This is over 200x faster overall execution.