Introduce generic CLI options framework, support "print -option val --"
And "bt -OPTS" too.  And probably others I no longer remember.

TAB completion fully supported.  That's pretty neat I think.  Give it
a go!  Try:

  (gdb) p -[TAB]
  -address         -array-indexes   -null-stop       -pretty          -static-members  -union
  -array           -elements        -object          -repeats         -symbol          -vtbl

  (gdb) bt -[TAB]
  -entry-values         -frame-arguments      -full                 -hide                 -no-filters           -raw-frame-arguments

Note that:

 - you can shorten option names, as long as unambiguous.
 - For boolean options, 0/1 stand for off/on.
 - For boolean options, "true" is implied.

So these are all equivalent:

 (gdb) print -object on -static-members off -pretty on -- foo
 (gdb) print -object -static-members off -pretty -- foo
 (gdb) print -object -static-members 0 -pretty -- foo
 (gdb) print -o -st 0 -p -- foo

And so are these:

 (gdb) bt -entry-values both
 (gdb) bt -e b

"--" explicitly terminates options.  This is necessary because some
expressions may start with "-".  E.g., if you have a variable called
"object" and you want to print its negative: "-object":

   (gdb) print -- -object

That's a minor potential script breakage, but I think it's well worth
it.  FIXME -- no longer true, because print now _requires_ --.

Note that the code is organized such that some of the options and the
"set/show" commands code is shared.  In particular, the "print"
options and the corresponding "set print" commands are defined with
the same structures.  Same thing for some of the "bt" options.

I think this is a better approach than trying to cram a bunch of
unrelated settings under a single global one-character namespace.

Still missing:  (warning, this is stale now)

 - run it by upstream, as an RFC.
 - comments.  mention why polymorphism isn't done with virtual methods.
 - convert more commands.
 - not integrated with the new "thread apply" "vqs" flags yet, which
   had gone in after this was originally written.
 - write tests.
 - write docs.